requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
pillow==10.3.0
PyJWT==2.1.0
python-dotenv==1.0.0
